Move mount/umount related actions to initramfs or inittab
This gets rid of some errors during shutdown and cleans up the
boot and shutdown sequence.

Unmounting the SD card still fails when the root file system is
an image file on the SD card (because the root fs is mounted rw).
Will be fixed later by mounting the root fs read-only if from an
image file.