Get the firmware version number from git if possible
If the firmware is built from a git checkout, amend the version numer in /etc/fw-ver.txt with output from 'git describe'. This allows to identify the exact commit from which the firmware has been built, which should be helpful for support.

If the build dir is not under git control (e.g. because the firmware has been built from a source tarball), use the base version number from the board config.