update Fork (#1)
* Enable TLS/SSL for python

* Get the firmware version number from git if possible

If the firmware is built from a git checkout, amend the version numer in /etc/fw-ver.txt with output from 'git describe'. This allows to identify the exact commit from which the firmware has been built, which should be helpful for support.

If the build dir is not under git control (e.g. because the firmware has been built from a source tarball), use the base version number from the board config.

* Use /etc/ftc_logo.png as screenshot if none exists in /tmp

* simplified desfault screenshot usage

* Another tiny screenshot adjustment

* Attempt to fix robopro offline mode

* Write downloaded ROBOPro files to SD card

* Avoid confusion with SymLink to rootfs.squashfs

The current build process creates a file rootfs.squashfs and a SymLink to this file named rootfs.img. 
rootfs.img is the file required for the TXT but due to the symlink cannot easily be copied to the FAT-formatted SD-Card.

As the rootfs.squashfs-File is (currently?) only used as link target but for no other purposes, the proposed solution is to rename rootfs.squashfs to rootfs.img, thereby making rootfs.img copyable as the other required files.