Add navigation menu.
The navigation is automatically generated from the page structure.

The text for navigation links is taken from the page title, and
defaults to the page URL if a page has no title.

Navigation preserves the currently selected language if possible, but
falls back to a foreign language if a linked page is not available in
the currently selected language.